+++ packages/ai-runtime/src/core/executor/base.ts
@@ -3,17 +3,20 @@ import { hash } from 'crypto';
 import type { DatasetObject } from '@blocklet/dataset-sdk/types';
 import { memoize } from '@blocklet/quickjs';
 import { call } from '@blocklet/sdk/lib/component';
-import { logger } from '@blocklet/sdk/lib/config';
+import config, { logger } from '@blocklet/sdk/lib/config';
 import Joi from 'joi';
 import jsonStableStringify from 'json-stable-stringify';
 import isEmpty from 'lodash/isEmpty';
 import isNil from 'lodash/isNil';
+import pick from 'lodash/pick';
 import toLower from 'lodash/toLower';
 
 import { AIGNE_RUNTIME_COMPONENT_DID } from '../../constants';
 import {
   AssistantResponseType,
   ExecutionPhase,
+  OutputVariable,
+  Parameter,
   RuntimeOutputProfile,
   RuntimeOutputVariable,
   Variable,
@@ -23,6 +26,7 @@ import {
 } from '../../types';
 import { isNonNullable } from '../../utils/is-non-nullable';
 import { CallAI, CallAIImage, GetAgent, GetAgentResult, RunAssistantCallback } from '../assistant/type';
+import { issueVC } from '../libs/blocklet/vc';
 import { HISTORY_API_ID, KNOWLEDGE_API_ID, MEMORY_API_ID, getBlockletAgent } from '../utils/get-blocklet-agent';
 import { renderMessage } from '../utils/render-message';
 import { nextTaskId } from '../utils/task-id';
@@ -254,7 +258,7 @@ export abstract class AgentExecutorBase<T> {
 
     if (result === undefined) {
       const outputs = await this.process({ inputs: this.finalInputs });
-      result = await this.validateOutputs({ inputs: this.finalInputs, outputs });
+      result = await this.validateOutputs({ inputs: this.finalInputs, outputs, processCallAgentOutputs: true });
 
       // set cache if needed
       if (!isEmpty(result) && agent.cache?.enable && agent.identity) {
@@ -322,6 +326,7 @@ export abstract class AgentExecutorBase<T> {
         messageId: this.context.messageId,
         clientTime: this.context.clientTime,
         user: this.context.user,
+        env: pick(config.env, 'appId', 'appName', 'appDescription', 'appUrl'),
       },
       $storage: {
         async getItem(key: string, { scope = 'global', onlyOne }: { scope?: VariableScope; onlyOne?: boolean } = {}) {
@@ -346,6 +351,13 @@ export abstract class AgentExecutorBase<T> {
             agentId: agentId || agent.identity.agentId,
             cacheKey: key,
           });
+        },
+      },
+      $blocklet: {
+        issueVC: (args: Omit<Parameters<typeof issueVC>[0], 'userDid'>) => {
+          const userDid = executor.context.user?.did;
+          if (!userDid) throw new Error('Issue VC requires user did');
+          return issueVC({ ...args, userDid });
         },
       },
     };
@@ -681,10 +693,12 @@ export abstract class AgentExecutorBase<T> {
     inputs,
     outputs,
     partial,
+    processCallAgentOutputs,
   }: {
     inputs?: { [key: string]: any };
     outputs?: { [key: string]: any };
     partial?: boolean;
+    processCallAgentOutputs?: boolean;
   }) {
     const { agent } = this;
 
@@ -695,10 +709,11 @@ export abstract class AgentExecutorBase<T> {
     const outputVariables = (agent.outputVariables ?? []).filter((i) => !i.hidden);
 
     const outputInputs = outputVariables.reduce((res, output) => {
-      const input =
-        output.from?.type === 'input'
-          ? agent.parameters?.find((input) => input.id === output.from?.id && !input.hidden)
-          : undefined;
+      let input: Parameter | undefined;
+      if (output.from?.type === 'input') {
+        const fromId = output.from.id;
+        input = agent.parameters?.find((input) => input.id === fromId && !input.hidden);
+      }
 
       if (input?.key && output.name) {
         const val = inputs?.[input.key];
@@ -708,7 +723,82 @@ export abstract class AgentExecutorBase<T> {
       return res;
     }, {});
 
-    return joiSchema.validateAsync({ ...outputs, ...outputInputs }, { stripUnknown: true });
+    const result = await joiSchema.validateAsync({ ...outputs, ...outputInputs }, { stripUnknown: true });
+
+    if (processCallAgentOutputs && this.agent.outputVariables) {
+      const callAgentOutputs = this.agent.outputVariables.filter(
+        (i): i is typeof i & { name: string; from: { type: 'callAgent'; callAgent: { agentId: string } } } =>
+          i.from?.type === 'callAgent' && !!i.from.callAgent?.agentId && !!i.name
+      );
+      const templateOutputs = this.agent.outputVariables.filter(
+        (i): i is typeof i & Required<Pick<typeof i, 'name' | 'valueTemplate'>> => !!i.valueTemplate?.trim() && !!i.name
+      );
+
+      const isOutputActive = async (output: OutputVariable) => {
+        if (!output.activeWhen?.trim()) return true;
+
+        return Joi.boolean().validate(
+          await renderMessage(output.activeWhen, { ...inputs, ...outputs, ...result }, { stringify: false })
+        ).value;
+      };
+
+      const v = Object.fromEntries(
+        (
+          await Promise.all(
+            callAgentOutputs.map(async (i) => {
+              if (!(await isOutputActive(i))) return null;
+
+              return [
+                i.name,
+                await (async () => {
+                  if (!this.agent.identity) throw new Error('Agent identity not found');
+
+                  const toolAgent = await this.context.getAgent({
+                    blockletDid: i.from.callAgent.blockletDid || this.agent.identity.blockletDid,
+                    projectId: i.from.callAgent.projectId || this.agent.identity.projectId,
+                    projectRef: this.agent.identity.projectRef,
+                    agentId: i.from.callAgent.agentId,
+                    working: this.agent.identity.working,
+                  });
+                  if (!toolAgent) throw new Error('Tool agent not found');
+
+                  const currentTaskId = nextTaskId();
+
+                  return await this.context
+                    .executor(toolAgent, {
+                      inputs: i.from.callAgent.inputs,
+                      variables: { ...inputs, ...outputs },
+                      taskId: currentTaskId,
+                      parentTaskId: this.options.taskId,
+                    })
+                    .execute();
+                })(),
+              ];
+            })
+          )
+        ).filter(isNonNullable)
+      );
+
+      Object.assign(result, v);
+
+      const v1 = Object.fromEntries(
+        (
+          await Promise.all(
+            templateOutputs.map(async (i) => {
+              if (!(await isOutputActive(i))) return null;
+              return [
+                i.name,
+                await renderMessage(i.valueTemplate!, { ...inputs, ...outputs, ...v }, { stringify: false }),
+              ];
+            })
+          )
+        ).filter(isNonNullable)
+      );
+
+      Object.assign(result, v1);
+    }
+
+    return result;
   }
 
   private async postProcessOutputs(agent: GetAgentResult, { outputs }: { outputs: { [key: string]: any } }) {