+++ packages/ai-runtime/src/core/executor/image-blender.ts
@@ -1,0 +1,46 @@
+import crypto from 'crypto';
+
+import { getComponentMountPoint } from '@blocklet/sdk/lib/component';
+import { env } from '@blocklet/sdk/lib/config';
+import { isNil } from 'lodash';
+import { joinURL, withQuery } from 'ufo';
+
+import { NFT_BLENDER_COMPONENT_DID } from '../../constants';
+import { ImageBlenderAssistant } from '../../types';
+import { renderMessage } from '../utils/render-message';
+import { AgentExecutorBase } from './base';
+
+export class ImageBlenderAgentExecutor extends AgentExecutorBase<ImageBlenderAssistant> {
+  override async process({ inputs }: { inputs: { [key: string]: any } }) {
+    const { agent } = this;
+
+    if (!agent.templateId) throw new Error(`Missing templateId in image blender agent ${agent.id}`);
+
+    const vars = { ...this.globalContext, ...inputs };
+
+    const dynamicData = Object.fromEntries(
+      await Promise.all(
+        Object.entries(agent.dynamicData ?? {}).map(async ([key, val]) => [
+          key,
+          isNil(val) || typeof val === 'string' ? await renderMessage(val.trim() || `{{${key}}}`, vars) : val,
+        ])
+      )
+    );
+
+    return {
+      $images: [
+        {
+          url: withQuery(
+            joinURL(
+              env.appUrl,
+              getComponentMountPoint(NFT_BLENDER_COMPONENT_DID),
+              '/api/templates/preview',
+              agent.templateId
+            ),
+            { sn: crypto.randomInt(1, 1000000), dynamicData: JSON.stringify(dynamicData) }
+          ),
+        },
+      ],
+    };
+  }
+}