+++ packages/ai-runtime/src/types/assistant/index.ts
@@ -41,7 +41,8 @@ export type Assistant =
   | ApiAssistant
   | FunctionAssistant
   | RouterAssistant
-  | CallAssistant;
+  | CallAssistant
+  | ImageBlenderAssistant;
 
 export type Role = 'system' | 'user' | 'assistant';
 
@@ -202,7 +203,20 @@ export type VariableType = VariableTypeBase &
 
 export type OutputVariable = VariableType & {
   variable?: { key: string; scope: VariableScope };
-  from?: { type: 'input' | 'output'; id: string };
+  valueTemplate?: string;
+  activeWhen?: string;
+  from?:
+    | { type?: undefined }
+    | { type: 'input' | 'output'; id?: string }
+    | {
+        type: 'callAgent';
+        callAgent?: {
+          blockletDid?: string;
+          projectId?: string;
+          agentId?: string;
+          inputs?: { [key: string]: any };
+        };
+      };
   appearance?: RuntimeOutputAppearance;
   initialValue?: RuntimeOutputVariablesSchema[RuntimeOutputVariable];
 };
@@ -232,6 +246,12 @@ export interface RouterAssistant extends AssistantBase {
   frequencyPenalty?: number;
   maxTokens?: number;
   model?: string;
+}
+
+export interface ImageBlenderAssistant extends AssistantBase {
+  type: 'imageBlender';
+  templateId?: string;
+  dynamicData?: { [key: string]: string };
 }
 
 export interface CallAssistant extends AssistantBase {